tryffle/pyproject.toml
[tool.poetry]
name = "tryffle"
version = "0.1.0"
description = "Finf the truffle you seek in a forest of document"
authors = ["Erwan Vinet <erwanvinet034@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "5"
djangorestframework = "^3.15.1"
PyPDF2 = "^3.0.1"
pillow = "^10.3.0"
psycopg2-binary = "^2.9.9"
celery = "^5.4.0"
redis = "^5.0.4"
rabbitmq = "^0.2.0"
pdf2image = "^1.17.0"
pyocr = "^0.8.5"
django-debug-toolbar = "^4.4.2"
django-vite = "^3.0.4"
inertia-django = "^0.6.0"
django-filter = "^24.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
